wip constrasints
